script/requirements.txt
azure-ai-textanalytics>=5.3.0, <6.0
beautifulsoup4>=4.12.3, <5.0
click>=8.1.7, <9.0
lxml>=5.2.2, <6.0
requests>=2.32.3, <3.0
